pyproject.toml
[project]
name = "spec2code"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.1",
    "jsonschema>=4.25.1",
    "networkx>=3.5",
    "pandas>=2.3.3",
    "pandera>=0.26.1",
    "pydantic>=2.12.3",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "deptry>=0.23.1",
    "mypy>=1.18.2",
    "pandas-stubs>=2.3.2.250926",
    "pytest>=8.4.2",
    "ruff>=0.9.0",
    "types-jsonschema>=4.25.1.20251009",
    "types-networkx>=3.5.0.20251001",
    "types-pyyaml>=6.0.12.20250915",
    "vulture>=2.14",
    "xenon>=0.9.3",
]

[tool.pytest.ini_options]
# testpaths = ["packages/tests"]
testpaths = ["spectool/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"
# Exclude problematic directories
norecursedirs = ["TransformFn", ".venv", ".git", "__pycache__", "apps", "output"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "ANN", "B", "S", "C90", "PLR", "RET", "SIM"]
ignore = ["S101", "B008", "ANN204", "B017", "ANN201"]
preview = true

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*.py" = ["ALL"]
"main.py" = ["PLR6301"]  # Fire CLI requires instance methods

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.pylint]
max-args = 7
max-branches = 12
max-statements = 50
max-returns = 6
max-nested-blocks = 4
max-locals = 15

[tool.pylint.main]
disable = "all"
enable = "too-many-lines"

[tool.pylint.format]
max-module-lines = 500

[tool.mypy]
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "fire"
ignore_missing_imports = true

[tool.deptry]
extend_exclude = ["apps", "output", "frontend", "packages/tests"]
ignore_notebooks = true

[tool.deptry.per_rule_ignores]
DEP001 = ["spec2code", "spectool"]  # Local package import (project itself)
